+++ src/Hl7.Fhir.Base/ElementModel/ElementNodeExtensions.cs
@@ -110,3 +110,5 @@ namespace Hl7.Fhir.ElementModel
             node as ScopedNode ?? new ScopedNode(node);
     }
 }
+
+#nullable restore