+++ src/Hl7.Fhir.Base/ElementModel/TypedElementParseExtensions.cs
@@ -8,6 +8,7 @@
 
 using Hl7.Fhir.Model;
 using Hl7.Fhir.Support.Poco;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -15,6 +16,7 @@ namespace Hl7.Fhir.ElementModel
 {
     public static class TypedElementParseExtensions
     {
+        #region ParseBindable
         /// <summary>
         /// Parses a bindeable type (code, Coding, CodeableConcept, Quantity, string, uri) into a FHIR coded datatype.
         /// Extensions will be parsed from the 'value' of the (simple) extension.
@@ -32,15 +34,21 @@ namespace Hl7.Fhir.ElementModel
         ///   'uri' => code
         /// </remarks>
         public static Element ParseBindable(this ITypedElement instance)
-        {
+#pragma warning disable CS0618 // Type or member is obsolete
+            => instance.ParseBindableInternal();
+#pragma warning restore CS0618 // Type or member is obsolete
 
+        [Obsolete("WARNING! Intended for internal API usage exclusively, interface IBaseElementNavigator can be changed in " +
+            "the near future.")]
+        public static Element ParseBindableInternal<T>(this IBaseElementNavigator<T> instance) where T : IBaseElementNavigator<T>
+        {
             return instance.InstanceType switch
             {
-                FhirTypeConstants.CODE => instance.ParsePrimitive<Code>(),
-                FhirTypeConstants.STRING => new Code(instance.ParsePrimitive<FhirString>()?.Value),
-                FhirTypeConstants.URI => new Code(instance.ParsePrimitive<FhirUri>()?.Value),
-                FhirTypeConstants.CODING => instance.ParseCoding(),
-                FhirTypeConstants.CODEABLE_CONCEPT => instance.ParseCodeableConcept(),
+                FhirTypeConstants.CODE => instance.ParsePrimitiveInternal<Code, T>(),
+                FhirTypeConstants.STRING => new Code(instance.ParsePrimitiveInternal<FhirString, T>()?.Value),
+                FhirTypeConstants.URI => new Code(instance.ParsePrimitiveInternal<FhirUri, T>()?.Value),
+                FhirTypeConstants.CODING => instance.ParseCodingInternal(),
+                FhirTypeConstants.CODEABLE_CONCEPT => instance.ParseCodeableConceptInternal(),
                 FhirTypeConstants.QUANTITY => parseQuantity(),
                 FhirTypeConstants.EXTENSION => parseExtension(),
                 _ => null,
@@ -49,20 +57,28 @@ namespace Hl7.Fhir.ElementModel
             Coding parseQuantity()
             {
                 var newCoding = new Coding();
-                var q = instance.ParseQuantity();
+                var q = instance.ParseQuantityInternal();
                 newCoding.Code = q.Code;
                 newCoding.System = q.System ?? "http://unitsofmeasure.org";
                 return newCoding;
             }
-
             Element parseExtension()
             {
                 var valueChild = instance.Children("value").FirstOrDefault();
-                return valueChild?.ParseBindable();
+                return valueChild?.ParseBindableInternal();
             }
         }
+        #endregion
 
+        #region ParseQuantity
         public static Model.Quantity ParseQuantity(this ITypedElement instance)
+#pragma warning disable CS0618 // Type or member is obsolete
+            => ParseQuantityInternal(instance);
+#pragma warning restore CS0618 // Type or member is obsolete
+
+        [Obsolete("WARNING! Intended for internal API usage exclusively, interface IBaseElementNavigator can be changed in " +
+            "the near future.")]
+        public static Quantity ParseQuantityInternal<T>(this IBaseElementNavigator<T> instance) where T : IBaseElementNavigator<T>
         {
             var newQuantity = new Quantity
             {
@@ -78,12 +94,31 @@ namespace Hl7.Fhir.ElementModel
 
             return newQuantity;
         }
+        #endregion
 
+        #region ParsePrimitive
         public static T ParsePrimitive<T>(this ITypedElement instance) where T : PrimitiveType, new()
-                    => new T() { ObjectValue = instance.Value };
+#pragma warning disable CS0618 // Type or member is obsolete
+            => ParsePrimitiveInternal<T, ITypedElement>(instance);
+#pragma warning restore CS0618 // Type or member is obsolete
 
+        [Obsolete("WARNING! Intended for internal API usage exclusively, interface IBaseElementNavigator can be changed in " +
+            "the near future.")]
+        public static T ParsePrimitiveInternal<T, U>(this IBaseElementNavigator<U> instance) where T : PrimitiveType, new() where U : IBaseElementNavigator<U>
+                    => new() { ObjectValue = instance.Value };
 
+        #endregion
+
+        #region ParseCoding
         public static Coding ParseCoding(this ITypedElement instance)
+#pragma warning disable CS0618 // Type or member is obsolete
+            => ParseCodingInternal(instance);
+#pragma warning restore CS0618 // Type or member is obsolete
+
+
+        [Obsolete("WARNING! Intended for internal API usage exclusively, interface IBaseElementNavigator can be changed in " +
+            "the near future.")]
+        public static Coding ParseCodingInternal<T>(this IBaseElementNavigator<T> instance) where T : IBaseElementNavigator<T>
         {
             return new Coding()
             {
@@ -94,8 +129,17 @@ namespace Hl7.Fhir.ElementModel
                 UserSelected = instance.Children("userSelected").SingleOrDefault()?.Value as bool?
             };
         }
+        #endregion
 
+        #region ParseResourceReference
         public static ResourceReference ParseResourceReference(this ITypedElement instance)
+#pragma warning disable CS0618 // Type or member is obsolete
+            => instance.ParseResourceReferenceInternal();
+#pragma warning restore CS0618 // Type or member is obsolete
+
+        [Obsolete("WARNING! Intended for internal API usage exclusively, interface IBaseElementNavigator can be changed in " +
+            "the near future.")]
+        public static ResourceReference ParseResourceReferenceInternal<T>(this IBaseElementNavigator<T> instance) where T : IBaseElementNavigator<T>
         {
             return new ResourceReference()
             {
@@ -103,17 +147,30 @@ namespace Hl7.Fhir.ElementModel
                 Display = instance.Children("display").GetString()
             };
         }
+        #endregion
 
+        #region ParseCodeableConcept
         public static CodeableConcept ParseCodeableConcept(this ITypedElement instance)
+#pragma warning disable CS0618 // Type or member is obsolete
+            => instance.ParseCodeableConceptInternal();
+#pragma warning restore CS0618 // Type or member is obsolete
+
+        [Obsolete("WARNING! Intended for internal API usage exclusively, interface IBaseElementNavigator can be changed in " +
+            "the near future.")]
+        public static CodeableConcept ParseCodeableConceptInternal<T>(this IBaseElementNavigator<T> instance) where T : IBaseElementNavigator<T>
         {
             return new CodeableConcept()
             {
                 Coding =
-                    instance.Children("coding").Select(codingNav => codingNav.ParseCoding()).ToList(),
+                    instance.Children("coding").Select(codingNav => codingNav.ParseCodingInternal()).ToList(),
                 Text = instance.Children("text").GetString()
             };
         }
+        #endregion
 
-        public static string GetString(this IEnumerable<ITypedElement> instance) => instance.SingleOrDefault()?.Value as string;
+#pragma warning disable CS0618 // Type or member is obsolete
+        public static string GetString<T>(this IEnumerable<T> instance) where T : IBaseElementNavigator<T>
+#pragma warning restore CS0618 // Type or member is obsolete
+            => instance.SingleOrDefault()?.Value as string;
     }
 }