+++ src/Hl7.Fhir.Base/Rest/BaseFhirClient.cs
@@ -938,6 +938,21 @@ namespace Hl7.Fhir.Rest
             return OperationAsync(operation, parameters, useGet).WaitResult();
         }
 
+        public virtual Task<Bundle?> ProcessMessageAsync(Bundle bundle, bool async = false, string? responseUrl = null, CancellationToken? ct = null)
+        {
+            if (bundle == null) throw new ArgumentNullException(nameof(bundle));
+
+            var tx = new TransactionBuilder(Endpoint).ProcessMessage(bundle, async, responseUrl).ToBundle();
+
+            return executeAsync<Bundle>(tx, new [] {HttpStatusCode.OK, HttpStatusCode.Accepted, HttpStatusCode.NoContent}, ct);
+        }
+
+        [Obsolete("Synchronous use of the FhirClient is strongly discouraged, use the asynchronous call instead.")]
+        public virtual Resource? ProcessMessage(Bundle messageBundle, bool async = false, string? responseUrl = null)
+        {
+            return ProcessMessageAsync(messageBundle, async, responseUrl).WaitResult();
+        }
+
         private Task<Resource?> internalOperationAsync(string operationName, string? type = null, string? id = null, string? vid = null,
             Parameters? parameters = null, bool useGet = false, CancellationToken? ct = null)
         {